Eishwar/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
qrcode[pil]==7.4.2
# python-barcode provides the 'barcode' module used by utils/barcode_generator.py
# Pin to a valid current release; extras 'images' pulls in Pillow for PNG output
python-barcode[images]==0.15.1
websockets==12.0
pytest==7.4.3
pytest-asyncio==0.21.1
alembic==1.13.0
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2